Removed useCommandsParameterized hook
